Move new unit tests into the SpaceSplitterTest.
