--- com.unity.hlod/Samples~/Assets/Tests/EditMode/SpaceSplitterSubTreeTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c5d15236ae905384ea684e5f2c93a11a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
